requirements.txt
Faker==8.1.2
psycopg2==2.8.6
python-dateutil==2.8.1
python-decouple==3.4
six==1.15.0
text-unidecode==1.3
